Add CVE & GHSA suppressions for ws package
Added new suppression entries in NVDSuppressions.xml for the `ws` package versions 8.14.1 and 8.16.0 to address specific CVEs and a GHSA advisory. These suppressions are based on the rationale that the server functionality of the `ws` package, which is the component likely affected by these vulnerabilities, is not utilized within our project. Specifically, suppressions for CVE-2024-37890 and GHSA-3h5v-q93c-6h6q were added for both versions, with an additional entry for `8.16.0` that duplicates the CVE-2024-37890 suppression and includes a `cpe` identifier for the `ws` package.