Disable test cases for exchange set format selection
Disabled test cases in `ess-ui-encselectionpage-nonAdmin.spec.ts`
and `ess-ui-encselectionpage.spec.ts` due to the unavailability
of the S57 format. Specifically, the test case for non-UKHO users
not seeing options on the 'Confirm exchange set content' screen
and the test case for UKHO users seeing options on the same screen
have been commented out.