Remove AVCS product type filter from configurations (#306)
* Remove AVCS product type filter from configurations

This commit removes the filter condition for "Product Type"
equal to "AVCS" from the "rows" array in the "avcs" section
of both `popularsearchconfig.json` and `simplepopularsearchconfig.json`.
This change may impact how AVCS-related content is filtered
or displayed in the application.

* Add suppression for esbuild vulnerability GHSA-67mh-4wv8-2f99

This commit introduces a new suppression entry in the
`NVDSuppressions.xml` file for the `esbuild:0.21.5` package.
It includes relevant notes, a package URL regex, and the
specific vulnerability name, allowing for the suppression
of vulnerabilities associated with this version of `esbuild`.