Update MSAL and GTM dependencies; resolve version issues (#349)
This is a Snyk initiated upgrade, the original PRs were too old to run reliably.

Upgraded `@azure/msal-angular` to `^3.1.0` and `@azure/msal-browser` to `^3.28.1` in `package.json` and `package-lock.json` for new features and fixes.

Downgraded `@azure/msal-browser` to `3.28.1` and `@azure/msal-common` to `14.16.0` in `package-lock.json` to address compatibility issues.

Updated `angular-google-tag-manager` to `^1.11.0` for improvements. Adjusted integrity hashes and resolved URLs for downgraded dependencies.